[ARROW-16904]: inverted conditions for MinMax
Based on multiple locations fixed in 85789a9, it seemed better to
simplify a few conditions so that updating `this->state` from `local` is
consolidated and more maintainable. This should make it easier to
understand that there is only a difference in how the local state is
updated when consuming input data and there is no difference in how the
aggregate's state is updated from the local state.